[REFACTOR] Order/Rank 인덱스 적용
