+++ src/main/java/com/github/gokid96/e_commerce/rank/domain/Rank.java
@@ -7,6 +7,7 @@ import jakarta.persistence.Enumerated;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
+import jakarta.persistence.Index;
 import jakarta.persistence.Table;
 import lombok.AccessLevel;
 import lombok.Builder;
@@ -17,7 +18,9 @@ import java.time.LocalDate;
 
 @Getter
 @Entity
-@Table(name = "product_rank")
+@Table(name = "product_rank", indexes = {
+        @Index(name = "idx_rank_type_date", columnList = "rank_type, rank_date")
+})
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
 public class Rank {
